Move some code around and disable MySQL
